Move ScrapeError to its own module
