+++ src/profile.rs
@@ -1,3 +1,4 @@
+use error::Result;
 use image::{PostImage, ProfileImage};
 use scrape;
 
@@ -37,7 +38,7 @@ impl From<scrape::JsonProfile> for Profile {
 }
 
 impl Profile {
-    pub fn get(username: &str) -> Result<Profile, scrape::ScrapeError> {
+    pub fn get(username: &str) -> Result<Profile> {
         let json_profile = scrape::scrape_profile(username)?;
         Ok(Profile::from(json_profile))
     }
